src/mdc/mdc-data-collector/data_collector/requirements.txt
flask
flask_restx
pymodbus
PyYaml
apscheduler
celery
requests
redis